% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{alternate_base_url}
\alias{alternate_base_url}
\title{Compute the alternate CHEF base URL}
\usage{
alternate_base_url(base_url)
}
\arguments{
\item{base_url}{Base API URL.}
}
\value{
Alternate base URL or \code{NULL}.
}
\description{
Compute the alternate CHEF base URL
}
\keyword{internal}
