% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingExport.R
\name{coerce_export_sheets}
\alias{coerce_export_sheets}
\title{Coerce input into exportable sheets}
\usage{
coerce_export_sheets(data, sheet = "Data")
}
\arguments{
\item{data}{Data frame, list of data frames, or flatten_submission_records list.}

\item{sheet}{Default sheet name for a single data frame.}
}
\value{
A named list of data frames.
}
\description{
Coerce input into exportable sheets
}
\keyword{internal}
