% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingExport.R
\name{coerce_list_columns}
\alias{coerce_list_columns}
\title{Coerce list columns to text}
\usage{
coerce_list_columns(df)
}
\arguments{
\item{df}{Data frame with possible list columns.}
}
\value{
Data frame with list columns converted to text.
}
\description{
Coerce list columns to text
}
\keyword{internal}
