% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{coerce_repeat_values}
\alias{coerce_repeat_values}
\title{Coerce repeated values to stable text}
\usage{
coerce_repeat_values(values)
}
\arguments{
\item{values}{Vector or list column values.}
}
\value{
A vector with list/data.frame values coerced to text.
}
\description{
Coerce repeated values to stable text
}
\keyword{internal}
