% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisHelpers.R
\name{coerce_time}
\alias{coerce_time}
\title{Coerce values to POSIXct}
\usage{
coerce_time(values, tz = "UTC")
}
\arguments{
\item{values}{Vector of values to parse.}

\item{tz}{Time zone for parsing.}
}
\value{
A POSIXct vector (possibly with \code{NA} values).
}
\description{
Tries to parse dates/times into POSIXct, falling back to Date parsing.
}
\keyword{internal}
