% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{collect_components}
\alias{collect_components}
\title{Collect components recursively}
\usage{
collect_components(
  components,
  parent_path = NULL,
  section = NULL,
  include = "input",
  expand_surveys = FALSE
)
}
\arguments{
\item{components}{Component list to traverse.}

\item{parent_path}{Current path prefix.}

\item{section}{Current section label.}

\item{include}{Which components to include.}

\item{expand_surveys}{Logical. Expand survey components into question rows.}
}
\value{
A list of row lists describing components.
}
\description{
Collect components recursively
}
\keyword{internal}
