% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisHelpers.R
\name{default_stopwords}
\alias{default_stopwords}
\title{Default stopword list}
\usage{
default_stopwords()
}
\value{
A character vector of stopwords.
}
\description{
Provides a small set of common English stopwords.
}
\keyword{internal}
