% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{default_yes_values}
\alias{default_yes_values}
\title{Default values treated as "yes"}
\usage{
default_yes_values()
}
\value{
A vector of values considered affirmative.
}
\description{
Default values treated as "yes"
}
\keyword{internal}
