% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{derive_prefix}
\alias{derive_prefix}
\title{Derive prefixes from column names}
\usage{
derive_prefix(names_vec, sep)
}
\arguments{
\item{names_vec}{Column names.}

\item{sep}{Separator between prefix and suffix.}
}
\value{
Prefixes or \code{NA} when no separator is present.
}
\description{
Derive prefixes from column names
}
\keyword{internal}
