% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{extract_child_components}
\alias{extract_child_components}
\title{Extract child components from a component}
\usage{
extract_child_components(comp)
}
\arguments{
\item{comp}{Component definition.}
}
\value{
A list of child components (possibly empty).
}
\description{
Extract child components from a component
}
\keyword{internal}
