% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{extract_flat_df}
\alias{extract_flat_df}
\title{Extract a flat data frame}
\usage{
extract_flat_df(x)
}
\arguments{
\item{x}{Data frame or flatten_submission_records list.}
}
\value{
A data frame of responses.
}
\description{
Extract a flat data frame
}
\keyword{internal}
