% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{extract_options}
\alias{extract_options}
\title{Extract selectable options from a component}
\usage{
extract_options(comp)
}
\arguments{
\item{comp}{Component definition.}
}
\value{
A character string of options, or \code{NULL} if unavailable.
}
\description{
Extract selectable options from a component
}
\keyword{internal}
