% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{fetch_form_schema}
\alias{fetch_form_schema}
\title{Fetch a form schema from the API}
\usage{
fetch_form_schema(
  base_url,
  form_id,
  api_key,
  version_id = NULL,
  tried_alt = FALSE
)
}
\arguments{
\item{base_url}{Base API URL.}

\item{form_id}{Form ID.}

\item{api_key}{API key.}

\item{version_id}{Optional version ID.}

\item{tried_alt}{Logical. Whether an alternate base URL has been tried.}
}
\value{
A schema list or \code{NULL}.
}
\description{
Fetch a form schema from the API
}
\keyword{internal}
