% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{first_non_empty}
\alias{first_non_empty}
\title{Return the first non-empty value}
\usage{
first_non_empty(...)
}
\arguments{
\item{...}{Candidate values.}
}
\value{
First non-empty character value or \code{NA_character_}.
}
\description{
Return the first non-empty value
}
\keyword{internal}
