% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenSubmissions.R
\name{FlattenSubmissions}
\alias{FlattenSubmissions}
\title{Flatten Submissions}
\usage{
FlattenSubmissions(x)
}
\arguments{
\item{x}{list with responses, typically obtained through the GetResponses() function.}
}
\value{
list with
\itemize{
\item \code{FlatResponses} which is the entire dataset obtained from GetResponses() but, where all the nested lists are flattened into a bidimensional data frame. nested lists will take their name by using their parent name as a prefix, then a "-" separator, and then their own name.
\item \code{ColumnNames} which is a data frame with column \code{Number} and column \code{Names}, listing all the newly created column names in \code{FlatResponses} and their numerical order. This output can then be passed to other FormIO functions for making easier to subset the dataset
}
}
\description{
This function will flatten submissions coming from a FormIO list that has nested elements
}
\examples{

x<-FoodTypes

# Nested Structure
head(x)

#Flattened Structure
xFlat<-FlattenSubmissions(x)

xFlat$FlatResponses ## Survey Output
xFlat$ColumnNames   ## New Columns Formed
}
