% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodTypes.R
\docType{data}
\name{FoodTypes}
\alias{FoodTypes}
\title{Survey Responses Dataset}
\format{
An object of class \code{data.frame} with 6 rows and 15 columns.
}
\source{
Based on standard structure from BC Public Service FormIO CHEF structure, processed and
cleaned using custom R scripts.
}
\usage{
data("FoodTypes")
}
\description{
A cleaned and structured dataset containing survey  mock responses collected from
individual respondents. Each row represents a unique respondent, identified
by a unique ID, and columns correspond to survey variables, metadata, and
derived indicators.
}
\details{
The dataset is intended for quantitative and mixed-methods analysis of
respondent-level patterns, including descriptive statistics, modeling, and
aggregation of survey-based indicators.
}
\keyword{datasets}
\keyword{social-science}
\keyword{survey}
