% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormIOr-package.R
\docType{package}
\name{FormIOr-package}
\alias{FormIOr}
\alias{FormIOr-package}
\title{FormIOr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
