% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuditLog.R
\name{is_audit_log_active}
\alias{is_audit_log_active}
\title{Check whether audit logging is active}
\usage{
is_audit_log_active()
}
\value{
\code{TRUE} if an audit log is active, otherwise \code{FALSE}.
}
\description{
This is most useful in scripts when you want to conditionally add a manual
note using \code{\link[=append_audit_log_entry]{append_audit_log_entry()}} only when logging is enabled.
}
\examples{
is_audit_log_active()
}
