% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{is_flat_list}
\alias{is_flat_list}
\title{Check whether input is a flatten_submission_records list}
\usage{
is_flat_list(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} is a list with \code{FlatResponses}, otherwise \code{FALSE}.
}
\description{
Check whether input is a flatten_submission_records list
}
\keyword{internal}
