% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{is_section_type}
\alias{is_section_type}
\title{Check if a type represents a section}
\usage{
is_section_type(type)
}
\arguments{
\item{type}{Component type string.}
}
\value{
\code{TRUE} if the type is a section-like layout.
}
\description{
Check if a type represents a section
}
\keyword{internal}
