% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{is_survey_type}
\alias{is_survey_type}
\title{Check if component type is a survey}
\usage{
is_survey_type(type)
}
\arguments{
\item{type}{Component type string.}
}
\value{
\code{TRUE} if the type is a survey component.
}
\description{
Check if component type is a survey
}
\keyword{internal}
