% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingExport.R
\name{last_segment}
\alias{last_segment}
\title{Extract the last segment of a key}
\usage{
last_segment(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Last segment after delimiters.
}
\description{
Extract the last segment of a key
}
\keyword{internal}
