% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudioAddin.R
\name{launch_formior_addin}
\alias{launch_formior_addin}
\title{Launch the FormIOr RStudio addin}
\usage{
launch_formior_addin()
}
\description{
Provides a minimal point-and-click interface for key FormIOr functions.
}
