% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{list_element_to_text}
\alias{list_element_to_text}
\title{Convert nested list elements to text}
\usage{
list_element_to_text(x)
}
\arguments{
\item{x}{List, data.frame, or atomic value.}
}
\value{
A character representation (or \code{NA_character_}).
}
\description{
Convert nested list elements to text
}
\keyword{internal}
