% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{normalize_checkbox_values}
\alias{normalize_checkbox_values}
\title{Normalize checkbox values for comparison}
\usage{
normalize_checkbox_values(values)
}
\arguments{
\item{values}{Vector of values.}
}
\value{
Normalized character vector.
}
\description{
Normalize checkbox values for comparison
}
\keyword{internal}
