% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportingExport.R
\name{normalize_details}
\alias{normalize_details}
\title{Normalize audit log details}
\usage{
normalize_details(details, n)
}
\arguments{
\item{details}{Details input.}

\item{n}{Expected length.}
}
\value{
Character vector of length \code{n}.
}
\description{
Normalize audit log details
}
\keyword{internal}
