% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{normalize_list}
\alias{normalize_list}
\title{Normalize list inputs}
\usage{
normalize_list(x)
}
\arguments{
\item{x}{List-like input.}
}
\value{
A list (possibly empty).
}
\description{
Normalize list inputs
}
\keyword{internal}
