% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{normalize_names}
\alias{normalize_names}
\title{Normalize column names}
\usage{
normalize_names(names_vec, style, make_unique, transliterate)
}
\arguments{
\item{names_vec}{Character vector of column names.}

\item{style}{Naming style ("snake", "lower", "upper", "title").}

\item{make_unique}{Logical. Ensure unique names.}

\item{transliterate}{Logical. Convert accents to ASCII.}
}
\value{
A character vector of normalized names.
}
\description{
Normalize column names
}
\keyword{internal}
