% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{pad_right}
\alias{pad_right}
\title{Pad a string to the right}
\usage{
pad_right(x, width)
}
\arguments{
\item{x}{Character input.}

\item{width}{Target display width.}
}
\value{
Right-padded string.
}
\description{
Pad a string to the right
}
\keyword{internal}
