% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{pick_compare_columns}
\alias{pick_compare_columns}
\title{Pick columns to compare duplicates}
\usage{
pick_compare_columns(
  df,
  id_col_name,
  compare_cols,
  quiet = FALSE,
  prompt = TRUE,
  default_cols = NULL
)
}
\arguments{
\item{df}{Data frame of responses.}

\item{id_col_name}{Submission ID column.}

\item{compare_cols}{Optional pre-selected columns.}

\item{quiet}{Logical. Suppress messages.}

\item{prompt}{Logical. Allow interactive prompting.}

\item{default_cols}{Default columns to use when prompting.}
}
\value{
Character vector of comparison column names.
}
\description{
Pick columns to compare duplicates
}
\keyword{internal}
