% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameCols.R
\name{RenameCols}
\alias{RenameCols}
\title{Rename columns from a FormIO response dataset}
\usage{
RenameCols(x, NamesDF = T, renameDF = T)
}
\arguments{
\item{x}{Flattened FormIO Dataset}

\item{NamesDF}{TRUE by default. Will return a dataframe, called 'renamedDF', with three columns: Number, OldNames and NewNames}

\item{renameDF}{TRUE by default. Will return the original data frame with the new columns assigned}
}
\value{
description list with:
\itemize{
\item data frame  called 'renamedDF', with three columns: Number, OldNames and NewNames
\item data frame with the original data frame with the new columns assigned
}
}
\description{
This fucntion will take a dataset object generated by FlattenSubmissions() and create an interfase to ask the user iteratively the name that they will assign to each of the columns from the object.
The function allows to return the new response dataset (NamesDF=T, by default), and return a dataframe with three columns: Number, OldNames and NewNames (renameDF=T)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# df<-FoodTypes

#Original Names
colnames(df)
RenameCols(df)


\dontshow{\}) # examplesIf}
}
