% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{try_fetch_schema}
\alias{try_fetch_schema}
\title{Attempt to fetch the schema from metadata}
\usage{
try_fetch_schema(form, version = "latest")
}
\arguments{
\item{form}{Metadata list.}

\item{version}{Version identifier.}
}
\value{
A schema list or \code{NULL}.
}
\description{
Attempt to fetch the schema from metadata
}
\keyword{internal}
