% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{try_fetch_schema_url}
\alias{try_fetch_schema_url}
\title{Try fetching a schema URL}
\usage{
try_fetch_schema_url(url, form_id, api_key)
}
\arguments{
\item{url}{Schema endpoint URL.}

\item{form_id}{Form ID.}

\item{api_key}{API key.}
}
\value{
A schema list or \code{NULL}.
}
\description{
Try fetching a schema URL
}
\keyword{internal}
