% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningSimplification.R
\name{update_column_names}
\alias{update_column_names}
\title{Update stored column name metadata}
\usage{
update_column_names(column_names_df, new_names)
}
\arguments{
\item{column_names_df}{Column metadata data.frame (optional).}

\item{new_names}{New column names.}
}
\value{
Updated metadata data.frame (or unchanged input).
}
\description{
Update stored column name metadata
}
\keyword{internal}
