% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataDocs.R
\name{values_equal}
\alias{values_equal}
\title{Compare two values for equality}
\usage{
values_equal(x, y)
}
\arguments{
\item{x}{First value.}

\item{y}{Second value.}
}
\value{
\code{TRUE} if the normalized values are equal.
}
\description{
Compare two values for equality
}
\keyword{internal}
